cleanup and comments
